api/pyproject.toml
[tool.poetry]
name = "employee_tracker"
version = "0.1.0"
description = ""
authors = ["Muhsin Kılıç <kmuhsinn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
channels = "^4.2.0"
channels-redis = "^4.2.1"
drf-yasg = "^1.21.8"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
django-grappelli = "^4.0.1"
flake8 = "^7.1.1"
black = "^24.10.0"
python-decouple = "^3.8"
django-filter = "^24.3"
django-debug-toolbar = "^4.4.6"
flower = "^2.0.1"
djangorestframework-simplejwt = "^5.3.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
django-cors-headers = "^4.6.0"
daphne = "^4.1.2"
logging = "^0.4.9.6"
ruff = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
